requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF
faiss-cpu
langchain_google_genai
sentence-transformers
numpy
